Documentation
This commit adds the Code Book for understanding the content of the data
as well as the following description of the of what the run_analysis.R
did to create each of the two output files.

Run analysis has the code fully documented, however, this is a
conceptual overview of what the program does.

1: load all libraries that are needed
2: define each of the fuction below
2.1: main- this is the core of the program and is responsible for each
of the major function calls
2.2: openFiles- this function imports a tab deliminated data file and
converts it to a data table
2.3 mergeDataFile- this function brings each of the 6 data files,
renames fields to prevent issues downstream, and column and row binds
the data as appropriate
2.4: appropriateDataIdentification: this function renames all columns in
the merged dataset and then selects the ones that are desired
2.5:relabeledData- this function joins the activity names into the
dataset
2.6: summarizeData- this function melts and summarizes the data from the
variable that creates the "All Data.txt" output to create the variable
that creates the "Summarized UCI HAR Dataset.txt" file.
2.7: output- this function writes the two outputs locally.
3: the main function is called and the program executed